Refactor Marathon. Fix looping bug
